+++ deploy/infrastructure/index.ts
@@ -1,9 +1,19 @@
 import * as aws from "@pulumi/aws";
 import * as vpc from "./vpc";
-import * as app from "./app_container";
 import * as alb from "./load_balancer";
 
 import { Cluster, Service } from "../components/cluster";
+import { Image } from "../components/image"
+
+const commonTags = {};
+
+const appImage = new Image('wci-hello', {
+  folder: "../app",
+  container: "wci/hello-world",
+  tag: "latest",
+  tags: { ...commonTags }
+})
+
 
 const cluster = new Cluster('wci-dev', {
   vpc: vpc.vpc,
@@ -17,7 +27,7 @@ const service = new Service('http', {
   taskExecutionRole: cluster.ecsTaskExecutionRole,
   vpc: vpc.vpc,
   alb: alb,
-  app: app,
+  app: { image: appImage.image, imageRef: appImage.image.ref },
   publicSubnets: vpc.publicSubnets,
   logGroup: logGroup,
   tags: { "magic": "foo" }
